% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{PrecisionCalc-package}
\alias{PrecisionCalc}
\alias{PrecisionCalc-package}
\title{PrecisionCalc: Sample Size and Distribution Analysis}
\description{
The PrecisionCalc package provides tools to compute required sample sizes
for various distributions, evaluate absolute or relative precision,
and visualize observed data with fitted distributions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ppoliqueta/PrecisionCalc}
}

}
\author{
\strong{Maintainer}: Paulo Paiva \email{paulo.paiva@gmail.com}

}
\keyword{internal}
