% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{relative_sample}
\alias{relative_sample}
\title{Sample Size Calculation for Relative Error}
\usage{
relative_sample(y, epsilon, conf = 0.95, Discrete = TRUE, N = NULL)
}
\arguments{
\item{y}{Numeric vector of counts.}

\item{epsilon_rel}{Desired relative error (0.1).}
}
\value{
Prints required sample sizes.
}
\description{
Computes sample sizes for both Discrete (Normal, Student-t, Poisson, and Negative Binomial) or Continuous (Normal, Lognormal and Gamma) distributions.
}
\examples{
y <- rpois(100, 5)
relative_sample(y, 0.1)
}
